Move modified_environ utility to flood_adapt.misc.utils and update imports
